Updated urllib3 due to security vulnerability
The urllib3 library before 1.24.2 for Python mishandles certain cases where the desired set of CA certificates is different from the OS store of CA certificates, which results in SSL connections succeeding in situations where a verification failure is a correct outcome. This is related to the use of the ssl_context, ca_certs, or ca_certs_dir argument.